MOSuite: R package for downstream multi-omics analysis

Designed for differential RNA-seq analysis or any data represented in a counts table.

Details

See the website for more information, documentation, and examples: https://ccbr.github.io/MOSuite

See also

Author

Maintainer: Kelly Sovacool kelly.sovacool@nih.gov (ORCID)

Authors:

Philip Homan philip.homan@nih.gov

Vishal Koparde vishal.koparde@nih.gov (ORCID)

Samantha Chill samantha.chill@nih.gov (ORCID)

Other contributors:

CCR Collaborative Bioinformatics Resource [copyright holder]