Volcano Plot - Summary [CC Produces one volcano plot for each tested contrast in the input DEG table.

It can be sorted by either fold change, t-statistic, or p-value. The returned dataset includes one row for each significant gene in each contrast, and contains columns from the DEG analysis of that contrast as well as columns useful to the Venn diagram template downstream.

Usage

plot_volcano_summary(
  moo_diff,
  feature_id_colname = NULL,
  signif_colname = "pval",
  signif_threshold = 0.05,
  change_threshold = 1,
  value_to_sort_the_output_dataset = "t-statistic",
  num_features_to_label = 30,
  add_features = FALSE,
  label_features = FALSE,
  custom_gene_list = "",
  default_label_color = "black",
  custom_label_color = "green3",
  label_x_adj = 0.2,
  label_y_adj = 0.2,
  line_thickness = 0.5,
  label_font_size = 4,
  label_font_type = 1,
  displace_feature_labels = FALSE,
  custom_gene_list_special_label_displacement = "",
  special_label_displacement_x_axis = 2,
  special_label_displacement_y_axis = 2,
  color_of_signif_threshold_line = "blue",
  color_of_non_significant_features = "black",
  color_of_logfold_change_threshold_line = "red",
  color_of_features_meeting_only_signif_threshold = "lightgoldenrod2",
  color_for_features_meeting_pvalue_and_foldchange_thresholds = "red",
  flip_vplot = FALSE,
  use_default_x_axis_limit = TRUE,
  x_axis_limit = 5,
  use_default_y_axis_limit = TRUE,
  y_axis_limit = 10,
  point_size = 2,
  add_deg_columns = c("FC", "logFC", "tstat", "pval", "adjpval"),
  graphics_device = grDevices::png,
  image_width = 15,
  image_height = 15,
  dpi = 300,
  use_default_grid_layout = TRUE,
  number_of_rows_in_grid_layout = 1,
  aspect_ratio = 0,
  plot_filename = "volcano.png",
  print_plots = options::opt("print_plots"),
  save_plots = options::opt("save_plots"),
  plots_subdir = "figures"
)

Arguments

moo_diff

Differential expression analysis result from one or more contrasts

feature_id_colname

The column from the counts data containing the Feature IDs (Usually Gene or Protein ID). This is usually the first column of your input Counts Matrix. Only columns of Text type from your input Counts Matrix will be available to select for this parameter. (Default: NULL - first column in the counts matrix will be used.)

signif_colname

column name of significance values (e.g., adjusted p-values or FDR). This column will be used to determine which points are considered significant in the volcano plot.

signif_threshold

Numeric value specifying the significance cutoff for p-values (i.e. filters on signif_colname)

change_threshold

Numeric value specifying the fold change cutoff for significance (i.e. filters on change_colname)

value_to_sort_the_output_dataset

How to sort the output dataset. Options are "fold-change" or "p-value".

num_features_to_label

Number of top features/genes to label in the volcano plot. Default is 30.

add_features

Add custom_gene_list To Labels. Set TRUE when you want to label a specific set of features (features) in the "custom_gene_list" parameter" IN ADDITION to the number of features you set in the "Number of Features to Label" parameter.

label_features

Select TRUE when you want to label ONLY a specific list of features(features) given in the "custom_gene_list" parameter.

custom_gene_list

Provide a list of features (comma separated) to be labeled on the volcano plot. You must toggle one of the following ON to see these labels: "Add features" or "Label Only My Feature List".

default_label_color

Set the color for the text used to add feature (gene) name labels to points.

custom_label_color

Set the color for the specific list of features (features) provided in the "Feature List" parameter.

label_x_adj

adjust position of the labels on the x-axis. Default: 0.2

label_y_adj

adjust position of the labels on the y-axis. Default: 0.2

line_thickness

Set the thickness of the lines in the plot. Default: 0.5

label_font_size

Set the font size of the labels. Default: 4

label_font_type

Set the font type of the labels. Default: 1

displace_feature_labels

Set to TRUE to displace gene labels. Default: FALSE. Set TRUE if you want to displace the feature (gene) label for a specific set of features. Make sure to use custom x- and y- limits and give sufficient space for displacement; otherwise other labels than the desired ones will appear displaced.

custom_gene_list_special_label_displacement

Provide a list of features (comma separated) for which you want special displacement of the feature label.

special_label_displacement_x_axis

Displacement of the feature label on the x-axis. Default: 2

special_label_displacement_y_axis

Displacement of the feature label on the y-axis. Default: 2

color_of_signif_threshold_line

Color of the significance threshold line. Default: "blue"

color_of_non_significant_features

Color of the non-significant features. Default: "black"

color_of_logfold_change_threshold_line

Color of the log fold change threshold line. Default: "red"

color_of_features_meeting_only_signif_threshold

Color of the features that meet only the significance threshold. Default: "lightgoldenrod2"

color_for_features_meeting_pvalue_and_foldchange_thresholds

Color of the features that meet both the p-value and fold change thresholds. Default: "red"

flip_vplot

Set to TRUE to flip the fold change values so that the volcano plot looks like a comparison was B-A. Default: FALSE

use_default_x_axis_limit

Set to TRUE to use the default x-axis limit. Default: TRUE

x_axis_limit

Custom x-axis limit. Default: c(-5, 5)

use_default_y_axis_limit

Set to TRUE to use the default y-axis limit. Default: TRUE

y_axis_limit

Custom y-axis limit. Default: c(0, 10)

point_size

Size of the points in the plot. Default: 1

add_deg_columns

Add additional columns from the DEG analysis to the output dataset. Default: FALSE

graphics_device

passed to ggsave(device). Default: grDevices::png

image_width

output image width in pixels - only used if save_plots is TRUE

image_height

output image height in pixels - only used if save_plots is TRUE

dpi

dots-per-inch of the output image (see ggsave()) - only used if save_plots is TRUE

use_default_grid_layout

Set to TRUE to use the default grid layout. Default: TRUE

number_of_rows_in_grid_layout

Number of rows in the grid layout. Default: 1

aspect_ratio

Aspect ratio of the output image. Default: 4/3

plot_filename

Filename for the output plot. Default: "volcano_plot.png"

print_plots

Whether to print plots during analysis (Defaults to FALSE, overwritable using option 'moo_print_plots' or environment variable 'MOO_PRINT_PLOTS')

save_plots

Whether to save plots to files during analysis (Defaults to FALSE, overwritable using option 'moo_save_plots' or environment variable 'MOO_SAVE_PLOTS')

plots_subdir

subdirectory in where plots will be saved if save_plots is TRUE